Work around Rubinius build script choking on RUBYLIB
Rubinius' `./configure` script is has a guard that aborts the build if
RUBYLIB environment variable is set. As an unfortunate consequence,
anybody using "rbenv-gem-rehash" plugin won't be able to build Rubinius.

This works around the issue by not writing to RUBYLIB if the name of the
script to be executed is `./configure`. Granted, this might affect more
projects than just Rubinius, but this assumes that a script named
`./configure` won't install any gems that need to be rehashed afterwards.

See rubinius/rubinius#2659